+++ .env
@@ -1,1 +1,1 @@
-REACT_APP_API_END_POINT=http://192.168.137.1:8081/dvdms/api/v1/
+REACT_APP_API_END_POINT=http://10.248.1.198:8081/dvdms/api/v1/
