+++ src/assets/styles/styles.css
@@ -10,6 +10,8 @@
  * Copyright 2011-2022 Twitter, Inc.
  * Licensed under MIT (https://github.com/twbs/bootstrap/blob/main/LICENSE)
  */
+
+
 :root {
   --bs-blue: #0d6efd;
   --bs-indigo: #6610f2;
@@ -8180,6 +8182,10 @@ textarea.form-control-lg {
   --bs-bg-opacity: 1;
   background-color: rgba(var(--bs-dark-rgb), var(--bs-bg-opacity)) !important;
 }
+.bg-dark2 {
+  --bs-bg-opacity: 1;
+  background-color: rgb(46, 51, 69) !important;
+}
 
 .bg-black {
   --bs-bg-opacity: 1;
@@ -10994,8 +11000,10 @@ body {
 .sb-sidenav .sb-sidenav-menu .nav .nav-link {
   display: flex;
   align-items: center;
-  padding-top: 0.75rem;
-  padding-bottom: 0.75rem;
+  /* padding-top: 0.75rem;
+  padding-bottom: 0.75rem; */
+  padding-top: 0.45rem;
+  padding-bottom: 0.45rem;
   position: relative;
 }
 .sb-sidenav .sb-sidenav-menu .nav .nav-link .sb-nav-link-icon {
@@ -11019,7 +11027,8 @@ body {
 }
 
 .sb-sidenav-dark {
-  background-color: #212529;
+  /* background-color: #212529; */
+  background-color: rgb(46, 51, 69);
   color: rgba(255, 255, 255, 0.5);
 }
 .sb-sidenav-dark .sb-sidenav-menu .sb-sidenav-menu-heading {
@@ -11240,4 +11249,48 @@ body {
   width: 20px !important;
   font-size: 0.75rem;
   border-radius: 0.375rem !important;
+}
+
+/* custom */
+html, body {
+  background-color: #f0f3f4;
+}
+body{
+  font-family: "Times New Roman", Times, serif;
+  font-size: 14px;
+  color: #58666e;
+  line-height: 1.42857143;
+}
+.sb-sidenav .sb-sidenav-menu .nav .nav-link {
+  height: 29px;
+  width: 100%;
+  line-height: 21px;
+  padding-left: 0px;
+}
+.nav .nav-link .sb-nav-link-icon, .sb-sidenav-menu .nav-link .sb-nav-link-icon {
+  margin-right: 0px;
+}
+.sb-sidenav .sb-sidenav-menu .nav .nav-link .sb-nav-link-icon i{
+  margin: 0px 12px 0px 16px;
+}
+.sb-sidenav{
+  padding: 15px 0px 30px;
+}
+.sb-sidenav-menu .nav .nav-link.active-me{
+  color: rgb(255, 255, 255);
+  background: rgb(0, 0, 0);
+}
+.sb-sidenav-menu .nav-link.active-me .sb-nav-link-icon i{
+  color: rgb(255, 255, 255);
+}
+.sb-sidenav-menu .nav .nav-link.active-me::before{
+  content: "";
+  width: 3px;
+  height: 100%;
+  border-radius: 0px 2px 2px 0px;
+  background: rgb(255, 255, 255);
+  position: absolute;
+}
+.sb-sidenav .sb-sidenav-menu .nav .sb-sidenav-menu-nested {
+  margin-left: 2.7rem;
 }
