--- src/dashboard/pages/demand/AnnualDemand.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-
-const AnnualDemand = () => {
-  return (
-    <>
-        <h1 className="mt-4">Demand</h1>
-        <ol className="breadcrumb mb-4">
-            <li className="breadcrumb-item active">Annual Demand Page</li>
-        </ol>
-    </>
-  )
-}
-
-export default AnnualDemand
