--- src/dashboard/pages/demand/DemandNotification.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-
-const DemandNotification = () => {
-  return (
-    <>
-        <h1 className="mt-4">Demand</h1>
-        <ol className="breadcrumb mb-4">
-            <li className="breadcrumb-item active">Notificarion</li>
-        </ol>
-    </>
-  )
-}
-
-export default DemandNotification
