+++ src/index.js
@@ -1,9 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import '@fortawesome/fontawesome-free/css/all.min.css';
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
-import "bootstrap/dist/css/bootstrap.min.css";
+//import "bootstrap/dist/css/bootstrap.min.css";
 import "bootstrap/dist/js/bootstrap.min.js";
 import { BrowserRouter } from "react-router-dom";
