+++ src/pages/layouts/Navbar.jsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom'
 const Navbar = () => {
   return (
     <>
-        <nav className="sb-topnav navbar navbar-expand navbar-dark bg-dark">
+        <nav className="sb-topnav navbar navbar-expand navbar-dark bg-dark2">
             {/* <!-- Navbar Brand--> */}
             <Link className="navbar-brand ps-3" to={'/'}>DVDMS</Link>
             {/* <!-- Sidebar Toggle--> */}
